Try tests with old structure and different runner in travis.
