+++ run.py
@@ -1,4 +1,4 @@
-from ufo.management_server import app
+import ufo
 
 if __name__ == '__main__':
-  app.run(debug=True, host='0.0.0.0')
+  ufo.app.run(debug=True, host='0.0.0.0')
