+++ ufo/models.py
@@ -1,4 +1,4 @@
-from management_server import db
+from . import db
 
 from Crypto.PublicKey import RSA
