+++ ufo/proxy_server.py
@@ -1,6 +1,6 @@
 """The module for handling proxy servers"""
 
-from management_server import app, db, setup_required
+from . import app, db, setup_required
 
 import flask
 import logging
