+++ ufo/setup.py
@@ -1,4 +1,4 @@
-from management_server import app, get_user_config
+from . import app, get_user_config
 
 import flask
 import httplib2
